+++ nostr-scrolls-macros/src/arg.rs
@@ -4,12 +4,12 @@
 use alloc::string::ToString;
 use proc_macro2::{Span, TokenStream};
 use quote::quote;
-use syn::{Attribute, FnArg, Ident, Pat, Path, Token, Type, spanned::Spanned};
+use syn::{Attribute, FnArg, Ident, Pat, Token, Type, TypePath, spanned::Spanned};
 
 /// Scroll argument attributes
 #[derive(Default)]
 struct ArgAtrrs {
-    from_ty: Option<Path>,
+    from_ty: Option<TypePath>,
 }
 
 impl ArgAtrrs {
@@ -37,7 +37,7 @@ impl ArgAtrrs {
     }
 
     fn parse_from(&mut self, attr: &Attribute) -> syn::Result<()> {
-        self.from_ty = Some(attr.parse_args::<Path>()?);
+        self.from_ty = Some(attr.parse_args::<TypePath>()?);
 
         Ok(())
     }
@@ -48,11 +48,13 @@ pub struct ScrollArg {
     attrs: ArgAtrrs,
     mutability: Option<Token![mut]>,
     name: Ident,
-    ty: Type,
+    ty: TypePath,
 }
 
-impl From<ScrollArg> for TokenStream {
-    fn from(arg: ScrollArg) -> Self {
+impl TryFrom<ScrollArg> for TokenStream {
+    type Error = syn::Error;
+
+    fn try_from(arg: ScrollArg) -> Result<Self, Self::Error> {
         let ScrollArg {
             attrs: ArgAtrrs { from_ty },
             mutability,
@@ -60,14 +62,24 @@ impl From<ScrollArg> for TokenStream {
             ty,
         } = arg;
 
-        if let Some(from_ty) = from_ty {
-            quote! {
+        match from_ty {
+            Some(from_ty)
+                if get_type_ident(&from_ty) == "Option" && get_type_ident(&ty) == "Option" =>
+            {
+                let inner_ty = get_option_inner(&ty)?;
+                let inner_from_ty = get_option_inner(&from_ty)?;
+
+                Ok(quote! {
+                    let #mutability #name: #ty = <#from_ty as nostr_scrolls::ReadParam>::read_param(ptr, &mut offset)
+                        .map(<#inner_ty as core::convert::From<#inner_from_ty>>::from);
+                })
+            }
+            Some(from_ty) => Ok(quote! {
                 let #mutability #name: #ty = <#ty as core::convert::From<#from_ty>>::from(<#from_ty as nostr_scrolls::ReadParam>::read_param(ptr, &mut offset));
-            }
-        } else {
-            quote! {
+            }),
+            None => Ok(quote! {
                 let #mutability #name: #ty = <#ty as nostr_scrolls::ReadParam>::read_param(ptr, &mut offset);
-            }
+            }),
         }
     }
 }
@@ -83,7 +95,7 @@ impl ScrollArg {
                 attrs: ArgAtrrs::parse(&arg_ident, &pat_type.attrs)?,
                 mutability: ident.mutability,
                 name: arg_ident,
-                ty: pat_type.ty.as_ref().clone(),
+                ty: get_type_path(&pat_type.ty)?,
             }));
         }
         Ok(None)
@@ -106,4 +118,30 @@ fn check_multiple_use_attr(
         return Err(err);
     }
     Ok(())
+}
+
+fn get_type_path(ty: &Type) -> syn::Result<TypePath> {
+    match ty {
+        Type::Path(ty_path) => Ok(ty_path.clone()),
+        _ => Err(syn::Error::new(ty.span(), "invalid type")),
+    }
+}
+
+fn get_type_ident(ty: &TypePath) -> alloc::string::String {
+    assert!(!ty.path.segments.is_empty());
+    ty.path.segments.last().unwrap().ident.to_string()
+}
+
+fn get_option_inner(ty: &TypePath) -> syn::Result<&TypePath> {
+    if let syn::PathArguments::AngleBracketed(a) =
+        &ty.path.segments.last().as_ref().unwrap().arguments
+        && let syn::GenericArgument::Type(option_ty) = a.args.first().unwrap()
+    {
+        if let Type::Path(type_path) = option_ty {
+            return Ok(type_path);
+        } else {
+            return Err(syn::Error::new(option_ty.span(), "invalid type"));
+        }
+    }
+    unreachable!()
 }