+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 
 - `ReadParam` for `isize` and `usize`
 - macros: `from` attr for parameters
+- macros: support `Option` in `from` attr
 
 ## v0.0.1 - 2026/04/10